MYTHOS-1158 - Removed LargeProto from proto in favor of long and decimal_places from CurrencyType

+++ saga-sdk-java/src/main/java/games/mythical/saga/sdk/client/executor/SagaCurrencyExecutor.java
@@ -2,11 +2,9 @@ package games.mythical.saga.sdk.client.executor;
 
 import games.mythical.saga.sdk.proto.common.currency.CurrencyState;
 
-import java.math.BigDecimal;
-
 public interface SagaCurrencyExecutor extends BaseSagaExecutor {
     void updateCurrency(String currencyId,
-                        BigDecimal amount,
+                        long amount,
                         String oauthId,
                         String traceId,
                         CurrencyState coinState) throws Exception;
